Update patch set 1
Patch Set 1:

Overall this change looks good, and we think it could resolve an issue that Jaikumar has been tracking.

It seems to be a known problem with DBUS
http://marc.info/?l=freedesktop-dbus&m=117323275932377&w=2

However we're not DBUS experts, the best place to get feedback to this patch would with upstream DBUS. They might very much appreciate this patch, or have better feedback for it.

This code is almost identical between 1.2.24 and 1.0.2, so I think you could submit the patch upstream as is.

We will then evaluate based on the feedback from upstream experts.

Also I included a couple of minor style comments, that would be good to fix before sending upstream.

Thanks,
Nick

Patch-set: 1