Add clarification to spec about UpdateActivationEnvironment
UpdateActivationEnvironment takes a a{ss}.  This means only
valid UTF-8 can be used.  Environment variables are normally
ascii, but in theory have no specific encoding to them.  This
means that certain valid environment variables can't be sent
to the bus for updating its activation environment.

This commit just adds a note to the spec explaining this
restriction.