Update patch set 2
Patch Set 2: Verified; No score

FWIW, the new patch set still builds fine.

Johan, are you planning to work with upstream dbus on this patch (or are you currently doing so?) - it'd be good to synchronize the efforts so that we don't keep this review open longer than necessary but also so that we don't lose track of it.

Patch-set: 2
Reviewer: Jean-Baptiste Queru <1000660@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0
Label: Verified=+1