Update patch set 2
Patch Set 2:

I assume that there are differences between our dbus and upstream 1.2, as the upstream change doesn't directly cherry-pick onto our trees.

Should we merge the change we have in Gerrit, with the understanding that we'll just use the upstream version if/when we merge from upstream dbus?

Patch-set: 2