[BT][DBUS]: reduced default timeout value
The issue occurs when a DBUS response is lost and the
synchronous call which was "waiting" for the response
is blocked until the timeout event is sent. This patch
reduces the value for this timeout in order to improve
user experience.

THE SIDE EFFECTS OF THIS PATCH SHOULD BE CAREFULLY
ASSESSED BEFORE VALIDATING IT AS THE CHANGE OF THE
VALUE IS SIGNIFICANT.

Change-Id: I08606bc370ce3b6e640e139ddc3376b17d7ad870
Author: Ovidiu Beldie <ovidiux.beldie@intel.com>
Signed-off-by: Ovidiu Beldie <ovidiux.beldie@intel.com>
Singed-off-by: Shuo Gao <shuo.gao@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 41065