dbus: add libstdc++ dependency.
Brillo "user" builds were failing to start dbus-daemon, and attempting
to start dbus-daemon manually gave a CANNOT_LINK_EXECUTABLE error
naming libstdc++.so.

I this this fixed it, but I have no idea if it's the right way to do
things. I've seen a variety of things in Android.mk files and don't
know which is best:
  * Use LOCAL_CXX_STL instead?
  * Statically link libstdc++ instead of dynamic?
  * Try to use libc++ instead? Does it matter which one?

To complicate things, the problem seems to be gone now, so either the
build system changed to fix this or another project just happened to
drag in libstdc++.so which incidentally fixed this. If it's the former
then this CL is no longer necessary.

Bug: http://b/24499744
Change-Id: I12422a0d2a6a223fc0c513f17328628c2638ee1e